+++ packet/packet_test.go
@@ -72,84 +72,88 @@ func TestParseMBAPHeader(t *testing.T) {
 	}
 }
 
-func TestIsLikeModbusTCP(t *testing.T) {
+func TestLooksLikeModbusTCP(t *testing.T) {
 	var testCases = []struct {
 		name                   string
 		when                   []byte
 		whenAllowUnsupportedFC bool
 		expectLength           int
-		expectLooksLike        LooksLikeType
+		expectError            string
 	}{
 		{
-			name:            "ok, full packet",
-			when:            []byte{0x01, 0x02, 0x00, 0x00, 0x00, 0x06, 0x10, 0x01, 0x00, 0x6B, 0x00, 0x03},
-			expectLength:    12,
-			expectLooksLike: LooksLikeTCPPacket,
+			name:         "ok, full packet",
+			when:         []byte{0x01, 0x02, 0x00, 0x00, 0x00, 0x06, 0x10, 0x01, 0x00, 0x6B, 0x00, 0x03},
+			expectLength: 12,
+			expectError:  "",
 		},
 		{
-			name:            "ok, fragment of packet",
-			when:            []byte{0x01, 0x02, 0x00, 0x00, 0x00, 0x06, 0x10, 0x01, 0x00},
-			expectLength:    12,
-			expectLooksLike: LooksLikeTCPPacket,
+			name:         "ok, fragment of packet",
+			when:         []byte{0x01, 0x02, 0x00, 0x00, 0x00, 0x06, 0x10, 0x01, 0x00},
+			expectLength: 12,
+			expectError:  "",
 		},
 		{
-			name:            "nok, ErrorResponseTCP (code=3)",
-			when:            []byte{0x81, 0x80, 0x0, 0x0, 0x0, 0x3, 0x1, 0x82, 0x3},
-			expectLength:    9,
-			expectLooksLike: UnsupportedFunctionCode,
+			name:         "nok, ErrorResponseTCP (code=3)",
+			when:         []byte{0x81, 0x80, 0x0, 0x0, 0x0, 0x3, 0x1, 0x82, 0x3},
+			expectLength: 9,
+			expectError:  "unsupported function code",
 		},
 		{
-			name:            "nok, too few bytes",
-			when:            []byte{0x01, 0x02, 0x00, 0x00, 0x00, 0x06, 0x10, 0x01},
-			expectLength:    0,
-			expectLooksLike: DataTooShort,
+			name:         "nok, too few bytes",
+			when:         []byte{0x01, 0x02, 0x00, 0x00, 0x00, 0x06, 0x10, 0x01},
+			expectLength: 0,
+			expectError:  "data is too short to be a Modbus TCP packet",
 		},
 		{
-			name:            "nok, invalid packet id, 1",
-			when:            []byte{0x01, 0x02, 0x01 /* 0x00 */, 0x00, 0x00, 0x06, 0x10, 0x01, 0x00, 0x6B, 0x00, 0x03},
-			expectLength:    0,
-			expectLooksLike: IsNotTPCPacket,
+			name:         "nok, invalid packet id, 1",
+			when:         []byte{0x01, 0x02, 0x01 /* 0x00 */, 0x00, 0x00, 0x06, 0x10, 0x01, 0x00, 0x6B, 0x00, 0x03},
+			expectLength: 0,
+			expectError:  "data does not like Modbus TCP packet",
 		},
 		{
-			name:            "nok, invalid packet id, 2",
-			when:            []byte{0x01, 0x02, 0x00, 0x01 /* 0x00 */, 0x00, 0x06, 0x10, 0x01, 0x00, 0x6B, 0x00, 0x03},
-			expectLength:    0,
-			expectLooksLike: IsNotTPCPacket,
+			name:         "nok, invalid packet id, 2",
+			when:         []byte{0x01, 0x02, 0x00, 0x01 /* 0x00 */, 0x00, 0x06, 0x10, 0x01, 0x00, 0x6B, 0x00, 0x03},
+			expectLength: 0,
+			expectError:  "data does not like Modbus TCP packet",
 		},
 		{
-			name:            "nok, pdu too short",
-			when:            []byte{0x01, 0x02, 0x00, 0x00, 0x00, 0x02 /* 0x04+ */, 0x10, 0x01, 0x00, 0x6B, 0x00, 0x03},
-			expectLength:    0,
-			expectLooksLike: IsNotTPCPacket,
+			name:         "nok, pdu too short",
+			when:         []byte{0x01, 0x02, 0x00, 0x00, 0x00, 0x02 /* 0x04+ */, 0x10, 0x01, 0x00, 0x6B, 0x00, 0x03},
+			expectLength: 0,
+			expectError:  "data does not like Modbus TCP packet",
 		},
 		{
-			name:            "nok, function code = 0",
-			when:            []byte{0x01, 0x02, 0x00, 0x00, 0x00, 0x06, 0x10, 0x00 /* 0x01 */, 0x00, 0x6B, 0x00, 0x03},
-			expectLength:    0,
-			expectLooksLike: IsNotTPCPacket,
+			name:         "nok, function code = 0",
+			when:         []byte{0x01, 0x02, 0x00, 0x00, 0x00, 0x06, 0x10, 0x00 /* 0x01 */, 0x00, 0x6B, 0x00, 0x03},
+			expectLength: 0,
+			expectError:  "data does not like Modbus TCP packet",
 		},
 		{
 			name:                   "ok, allow unsupported function code = 1F",
 			when:                   []byte{0x01, 0x02, 0x00, 0x00, 0x00, 0x06, 0x10, 0x1f /* 0x01 */, 0x00, 0x6B, 0x00, 0x03},
 			whenAllowUnsupportedFC: true,
 			expectLength:           12,
-			expectLooksLike:        LooksLikeTCPPacket,
+			expectError:            "",
 		},
 		{
 			name:                   "ok, unsupported function code = 1F",
 			when:                   []byte{0x01, 0x02, 0x00, 0x00, 0x00, 0x06, 0x10, 0x1f /* 0x01 */, 0x00, 0x6B, 0x00, 0x03},
 			whenAllowUnsupportedFC: false,
 			expectLength:           12,
-			expectLooksLike:        UnsupportedFunctionCode,
+			expectError:            "unsupported function code",
 		},
 	}
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			expectedLen, looksLike := IsLikeModbusTCP(tc.when, tc.whenAllowUnsupportedFC)
+			expectedLen, err := LooksLikeModbusTCP(tc.when, tc.whenAllowUnsupportedFC)
 
 			assert.Equal(t, tc.expectLength, expectedLen)
-			assert.Equal(t, tc.expectLooksLike, looksLike)
+			if tc.expectError != "" {
+				assert.EqualError(t, err, tc.expectError)
+			} else {
+				assert.NoError(t, err)
+			}
 		})
 	}
 }